KDP-3177 Allow Auth Via Keycloak
