+++ code-env/python/desc.json
@@ -2,5 +2,6 @@
   "acceptedPythonInterpreters": ["PYTHON37", "PYTHON38"],
   "forceConda": false,
   "installCorePackages": true,
+  "corePackagesSet": "PANDAS13",
   "installJupyterSupport": false
 }
