+++ code-env/python/spec/requirements.txt
@@ -1,1 +1,2 @@
-kdp-python-connector==0.30.0
+kdp-python-connector==0.37.6
+urllib3==1.25.3
